feat: implement user dropdown components for logged-in and anonymous users
